+++ packages/orq-rc/src/models/components/evaluatorresponseragas.ts
@@ -54,8 +54,8 @@ export const EvaluatorResponseRagas$inboundSchema: z.ZodType<
 > = z.object({
   _id: z.string(),
   description: z.string(),
-  created: z.string().default("2026-08-25T08:05:15.155Z"),
-  updated: z.string().default("2026-08-25T08:05:15.155Z"),
+  created: z.string().default("2026-08-25T10:39:50.840Z"),
+  updated: z.string().default("2026-08-25T10:39:50.840Z"),
   updated_by_id: z.nullable(z.string()).optional(),
   project_id: z.string().optional(),
   guardrail_config: z.any().optional(),