+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -3896,7 +3896,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01M0VZ7Q81C2XPYE165RYM8QAT"),
+  id: z.string().default("01M0W82S8A3SDA2W0CDP6PNDY2"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),