+++ packages/orq-rc/src/models/operations/streamrunagent.ts
@@ -3955,7 +3955,7 @@ export const AgentToolInputRunTools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   AgentToolInputRunTools
 > = z.object({
-  id: z.string().default("01M0VZ7Q981PBFYESGJ51S09BY"),
+  id: z.string().default("01M0W82S9927TRFAPYNBMGZ9H2"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() =>